Refactor Mapper class to extend ObjectMapper
This commit refactors the Mapper class to extend the ObjectMapper class from the com.fasterxml.jackson.databind package. It also adds a new method getInstance() to retrieve a singleton instance of the Mapper class. The disable(SerializationFeature.FAIL_ON_EMPTY_BEANS) method is called in the constructor to disable serialization failure on empty beans.